Crate cdrs

cdrs is a native Cassandra DB client written in Rust. It's under a hard development as of now.

Modules

client

consistency
The module contains Rust representation of Cassandra consistency levels.

error
This modules contains Cassandra's errors which server could respond to client.

frame
frame module contains general Frame functionality.

frame_query
Contains Query Frame related functionality.

frame_ready

frame_response

frame_result

frame_startup

types

value

Traits

AsByte
AsBytes should be used to convert a value into a single byte.

FromBytes
FromBytes should be used to parse an array of bytes into a structure.

FromCursor
FromCursor should be used to get parsed structure from an io:Cursor wich bound to an array of bytes.

FromSingleByte
FromSingleByte should be used to convert a single byte into a value. It is opposite to AsByte.

IntoBytes
IntoBytes should be used to convert a structure into array of bytes.