Module cdrs::error

This modules contains Cassandra's errors which server could respond to client.

Structs

AlreadyExistsError
The query attempted to create a keyspace or a table that was already existing. Read more...

CDRSError
CDRS error which could be returned by Cassandra server as a response. As it goes from the specification it contains an error code and an error message. Apart of those depending of type of error it could contain an additional information about an error. This additional information is represented by additional_info property which is ErrorKind.

FunctionFailureError
A (user defined) function failed during execution.

ReadFailureError
A non-timeout exception during a read request.

ReadTimeoutError
Timeout exception during a read request.

SimpleError
Is used if error does not contain any additional info.

UnavailableError
Additional info about unavailable exception

UnpreparedError
Can be thrown while a prepared statement tries to be executed if the provided prepared statement ID is not known by this host. Read more...

WriteFailureError
A non-timeout exception during a write request. Read more...

WriteTimeoutError
Timeout exception during a write request.

Enums

AdditionalErrorInfo
Additional error info in accordance to Cassandra protocol v4.

WriteType
Describes the type of the write that failed. Read more...

Type Definitions

Result
CDRS specific Result which contains a Frame in case of Ok and CDRSError if Err.