Enum cdrs::error::WriteType

pub enum WriteType {
    Simple,
    Batch,
    UnloggedBatch,
    Counter,
    BatchLog,
}

Describes the type of the write that failed. Read more...

Variants
Simple

The write was a non-batched non-counter write

Batch

The write was a (logged) batch write. If this type is received, it means the batch log has been successfully written

UnloggedBatch

The write was an unlogged batch. No batch log write has been attempted.

Counter

The write was a counter write (batched or not)

BatchLog

The failure occured during the write to the batch log when a (logged) batch write was requested.