Struct cdrs::error::CDRSError

pub struct CDRSError {
    pub error_code: CInt,
    pub message: CString,
    pub additional_info: AdditionalErrorInfo,
}

CDRS error which could be returned by Cassandra server as a response. As it goes from the specification it contains an error code and an error message. Apart of those depending of type of error it could contain an additional information about an error. This additional information is represented by additional_info property which is ErrorKind.

Fields
error_code: CInt

i32 that points to a type of error.

message: CString

Error message string.

additional_info: AdditionalErrorInfo

Additional information.