Struct cdrs::error::ReadFailureError

pub struct ReadFailureError {
    pub cl: Consistency,
    pub received: CInt,
    pub blockfor: CInt,
    pub num_failures: CInt,
    // some fields omitted
}

A non-timeout exception during a read request.

Fields
cl: Consistency

Consistency level of query.

received: CInt

i32 representing the number of nodes having acknowledged the request.

blockfor: CInt

i32 representing the number of replicas whose acknowledgement is required to achieve cl.

num_failures: CInt

Represents the number of nodes that experience a failure while executing the request.

Methods

impl ReadFailureError

fn replica_has_responded(&self) -> bool

Shows if replica has resonded to a query.