Struct cdrs::frame::frame_prepare::BodyReqPrepare

pub struct BodyReqPrepare { /* fields omitted */ }

Struct that represents a body of a frame of type prepare

Methods

impl BodyReqPrepare

fn new(query: String) -> BodyReqPrepare

Creates new body of a frame of type prepare that prepares query query.