Enum cdrs::frame::frame_result::ResResultBody

pub enum ResResultBody {
    Void(BodyResResultVoid),
    Rows(BodyResResultRows),
    SetKeyspace(BodyResResultSetKeyspace),
    Prepared(BodyResResultPrepared),
}

ResponseBody is a generalized enum that represents all types of responses. Each of enum option wraps related body type.

Variants
Void(BodyResResultVoid)

Void response body. It's an empty stuct.

Rows(BodyResResultRows)

Rows response body. It represents a body of response which contains rows.

SetKeyspace(BodyResResultSetKeyspace)

Set keyspace body. It represents a body of set_keyspace query and usually contains a name of just set namespace.

Prepared(BodyResResultPrepared)

Prepared response body.

Methods

impl ResResultBody

fn parse_body(cursor: &mut Cursor<Vec<u8>>, result_kind: ResultKind) -> ResResultBody

fn parse_body_from_cursor(cursor: &mut Cursor<Vec<u8>>, result_kind: ResultKind) -> ResResultBody

It retrieves ResResultBody from io::Cursor having knowledge about expected kind of result.