Struct cdrs::frame::Frame

pub struct Frame {
    pub version: Version,
    pub flags: Vec<Flag>,
    pub opcode: Opcode,
    pub stream: u64,
    pub body: Vec<u8>,
}

Fields
version: Version flags: Vec<Flag> opcode: Opcode stream: u64 body: Vec<u8>
Methods

impl Frame

fn new_req_auth_response(bytes: Vec<u8>) -> Frame

Creates new frame of type AuthResponse.

impl Frame

fn new_req_execute(id: CBytesShort, query_parameters: ParamsReqQuery) -> Frame

Note: This function should be used internally for building query request frames.

impl Frame

fn new_req_options() -> Frame

Creates new frame of type options.

impl Frame

fn new_req_prepare(query: String) -> Frame

Note: This function should be used internally for building query request frames.

impl Frame

fn new_req_query<'a>(query: String, consistency: Consistency, values: Option<Vec<Value>>, with_names: Option<bool>, page_size: Option<i32>, paging_state: Option<CBytes>, serial_consistency: Option<Consistency>, timestamp: Option<i64>) -> Frame

Note: This function should be used internally for building query request frames.

impl Frame

fn new_req_startup(compression: Option<String>) -> Frame

Creates new frame of type startup.

impl Frame

fn get_body(&self) -> ResponseBody

fn encode_with(self, compressor: Compression) -> Result<Vec<u8>>