Module cdrs::frame_result

Structs

BodyResResultRows
Structure that represents result of type rows.

BodyResResultSetKeyspace
It represents set keyspace result body. Body contains keyspace name.

BodyResResultVoid
Body of a response of type Void

CUdt
User defined type. Read more...

ColSpec
Single column specification.

ColTypeOption
Cassandra option that represent column type.

RowsMetadata
Rows metadata.

Enums

ColType
Cassandra data types which clould be returned by a server.

ColTypeOptionValue
Enum that represents all possible types of value of ColTypeOption.

ResResultBody
ResponseBody is a generalized enum that represents all types of responses. Each of enum option wraps related body type.

ResultKind
ResultKind is enum which represents types of result.

RowsMetadataFlag
Enum that represent a set of possible row metadata flags that could be set.