Trait cdrs::IntoBytes

pub trait IntoBytes {
    fn into_cbytes(&self) -> Vec<u8>;
}

IntoBytes should be used to convert a structure into array of bytes.

Required Methods

fn into_cbytes(&self) -> Vec<u8>

It should convert a struct into an array of bytes.